Reorganized tests into one folder and added a snapshot test for dropdown
